trainingApp could not be installed universally
To make trainingApp installable in any client no matter where the additional database gets installed, the additional database referenced in the screen and msg.ma files need to have the same screen number as the additional database referenced in the schm file (where it was originally exported from).

In out images during test installation, this only works since the database always gets installed in 25 and that was referenced in the screen and msg.ma files, but it never got updated during dependent update.